// src/app/models/hojacampo.models.ts
export class HojacampoModel {
    dias_det: string;
    des_det: string;
    prepa_suelo: string;
    prepa_semi: string;
    siembra: string;
    resiembra: string;
    deshierbe: string;
    aporque: string;
    lim_deshoje: string;
    elabo_abosoli: string;
    aplica_abono: string;
    contro_plaga: string;
    contro_enfer: string;
    mant_finca: string;
    cosecha: string;
    acarreo_transp: string;
    descarga: string;
    calibra_mante: string;
    molienda: string;
    filtrado: string;
    melada: string;
    clarificada: string;
    punteo: string;
    batido: string;
    tamizado: string;
    empacado: string;
    codificado: string;
    lim_modulo: string;
    desinf_modulo: string;
    regis_venta: string;
    llenado_conta: string;
    capacitacion: string;
    comercializacion: string;
    subjornal: number;
    costojornal: number;
    costomanoobra: number;
    manoobrafam: number;
    totmanoobra: number;
    combustible: number;
    transporte: number;
    mantenimiento: number;
    cana: string;
    totallaborescult: number;
    trapicheyhorno: string;
    tinasyutencillos: number;
    infraestructura: number;
    totalequiposymaquinaria: number;
}


export interface Hojacampo {
    dias_det: string;
    des_det: string;
    prepa_suelo: string;
    prepa_semi: string;
    siembra: string;
    resiembra: string;
    deshierbe: string;
    aporque: string;
    lim_deshoje: string;
    elabo_abosoli: string;
    aplica_abono: string;
    contro_plaga: string;
    contro_enfer: string;
    mant_finca: string;
    cosecha: string;
    acarreo_transp: string;
    descarga: string;
    calibra_mante: string;
    molienda: string;
    filtrado: string;
    melada: string;
    clarificada: string;
    punteo: string;
    batido: string;
    tamizado: string;
    empacado: string;
    codificado: string;
    lim_modulo: string;
    desinf_modulo: string;
    regis_venta: string;
    llenado_conta: string;
    capacitacion: string;
    comercializacion: string;
    subjornal: number;
    costojornal: number;
    costomanoobra: number;
    manoobrafam: number;
    totmanoobra: number;
    combustible: number;
    transporte: number;
    mantenimiento: number;
    cana: string;
    totallaborescult: number;
    trapicheyhorno: string;
    tinasyutencillos: number;
    infraestructura: number;
    totalequiposymaquinaria: number;
}